requirements.txt
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
pytz==2023.3
schedule==1.2.0 